+++ frontend/src/contact-us-page/ContactInfo.tsx
@@ -1,0 +1,24 @@
+import React from 'react'
+
+const ContactInfo = () => {
+  return (
+    <section className="contact-info">
+      <div className="info-block">
+        <h2>Contact Information</h2>
+        <p>
+          <strong>
+            NJ Department of Labor and Workforce Development
+          </strong>
+        </p>
+        <p>
+          Center for Occupational Employment Information (COEI)
+        </p>
+        <p>
+          PO Box 057, 5th Floor, Trenton, New Jersey 08625-0057
+        </p>
+      </div>
+    </section>
+  )
+}
+
+export default ContactInfo
