+++ frontend/src/contact-us-page/ContactSuccess.tsx
@@ -1,0 +1,15 @@
+import ResetFormBtn from "./ResetFormBtn"
+
+const ContactSuccess = () => {
+  return (
+    <section>
+      <h2>Success!</h2>
+      <p>
+        Your message has been sent and we will do our best to respond within 3-5 business days. Be sure to check your spam folder for any communications.
+      </p>
+      <ResetFormBtn />
+    </section>
+  )
+}
+
+export default ContactSuccess
