+++ frontend/src/contact-us-page/ResetFormBtn.tsx
@@ -1,0 +1,13 @@
+
+
+const ResetFormBtn = () => {
+  return (
+    <div className="button-container">
+      <button className="usa-button">
+        Reset Form 
+      </button>
+    </div>
+  )
+}
+
+export default ResetFormBtn
